+++ TechxManagementApi/Models/Projects/DeleteProjectRequest.cs
@@ -4,7 +4,7 @@ namespace TechxManagementApi.Models.Projects
 {
 	public class DeleteProjectRequest
 	{
-        public long ProjectId { get; set; }
+        public int ProjectId { get; set; }
 	}
 }
