+++ TechxManagementApi/Models/Teams/UpdateTeamNameRequest.cs
@@ -4,7 +4,7 @@ namespace TechxManagementApi.Models.Teams
 {
 	public class UpdateTeamNameRequest
 	{
-		public long TeamId { get; set; }
+		public int TeamId { get; set; }
         public string TeamName { get; set; }
 	}
 }
